Fixed double title bar when mediumweight popups are used.

+++ core/src/main/java/com/github/weisj/darklaf/ui/rootpane/DarkRootPaneUI.java
@@ -47,6 +47,7 @@ import java.beans.PropertyChangeEvent;
 public class DarkRootPaneUI extends BasicRootPaneUI implements HierarchyListener {
 
     protected static final String KEY_PREFIX = "JRootPane.";
+    public static final String KEY_IS_MEDIUM_WEIGHT_POPUP_ROOT = "mediumWeightPopupRoot";
     public static final String KEY_IS_POPUP = KEY_PREFIX + "isPopup";
     public static final String KEY_IS_TOOLTIP = KEY_PREFIX + "isToolTip";
     private Window window;
@@ -135,8 +136,8 @@ public class DarkRootPaneUI extends BasicRootPaneUI implements HierarchyListener
 
     private void installClientDecorations(final JRootPane root) {
         int style = windowDecorationsStyle < 0 ? root.getWindowDecorationStyle() : windowDecorationsStyle;
-        CustomTitlePane titlePane = Decorations.createTitlePane(root, style);
         updateWindow(root.getParent());
+        CustomTitlePane titlePane = Decorations.createTitlePane(root, style, window);
         setWindowDecorated();
         installLayout(root);
         setTitlePane(root, titlePane);
@@ -204,7 +205,9 @@ public class DarkRootPaneUI extends BasicRootPaneUI implements HierarchyListener
             });
         }
         if (e.getChangeFlags() == HierarchyEvent.PARENT_CHANGED) {
-            updateClientDecoration();
+            if (DarkUIUtil.getWindow(rootPane) != window) {
+                updateClientDecoration();
+            }
         }
     }
 
@@ -213,9 +216,11 @@ public class DarkRootPaneUI extends BasicRootPaneUI implements HierarchyListener
     }
 
     protected void updateClientDecoration() {
-        uninstallClientDecorations(rootPane);
-        if (Decorations.isCustomDecorationSupported() && !isPopup(rootPane)) {
-            installClientDecorations(rootPane);
+        if (!Boolean.TRUE.equals(rootPane.getClientProperty(KEY_IS_MEDIUM_WEIGHT_POPUP_ROOT))) {
+            uninstallClientDecorations(rootPane);
+            if (Decorations.isCustomDecorationSupported() && !isPopup(rootPane)) {
+                installClientDecorations(rootPane);
+            }
         }
     }
